refactor workspace manager into modular components
