+++ V2_COMPLIANCE_PROGRESS_TRACKER.md
@@ -487,6 +487,9 @@ Code quality, organization, and maintainability over strict line count limits
     - Split monolithic tests into game setup, agent, scenario, and validation modules
     - Added orchestrator for targeted execution
 
+17. ✅ **MAJOR-006**: Workspace Manager Modularization
+    - Split monolithic workspace manager into creator, coordinator, validator, and config modules
+
 ### Impact:
 - **Total lines reduced**: 14,962 → 3,466 lines (76.8% reduction)
 - **SRP compliance achieved** across all completed modules
